% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickem.R
\name{calculate_points}
\alias{calculate_points}
\title{Calculates points for each scenario}
\usage{
calculate_points(pickem, outcomes, losers = NULL)
}
\arguments{
\item{losers}{}
}
\description{
Calculates points for each scenario
}
