% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickem.R
\name{generate_outcomes}
\alias{generate_outcomes}
\title{Generate all possible outcomes for a given week and assign scenario
probability}
\usage{
generate_outcomes(probs, week)
}
\arguments{
\item{week}{}
}
\value{
data.table
}
\description{
Generate all possible outcomes for a given week and assign scenario
probability
}
