% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickem.R
\name{read_pickem}
\alias{read_pickem}
\title{Read Jeremy's Pick'Em excel file}
\usage{
read_pickem(filename, week, probs)
}
\arguments{
\item{filename}{path to the excel file}

\item{week}{week that data should be pulled from}

\item{probs}{}
}
\value{
data.table
}
\description{
Read Jeremy's Pick'Em excel file
}
